requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Vector Database
sentence-transformers==2.2.2
faiss-cpu==1.13.2

# Data Processing
numpy==1.24.3

# Mistral AI
mistralai>=1.2.0

# Document Processing
python-docx==1.1.0
PyPDF2>=3.0.0
